Ton of code cleanup activities.  Embraced newer APIs, fixed some uses of deprecated APIs.  Couple of bugs along the way.
git-svn-id: http://svn.igniterealtime.org/svn/repos/spark/trunk@9549 b35dd754-fafc-0310-a699-88a17e54d16e

+++ src/java/org/jivesoftware/sparkimpl/updater/SparkVersion.java
@@ -123,7 +123,7 @@ public class SparkVersion extends IQ {
                     }
                     else if (parser.getName().equals("updatedTime")) {
                         Long time = new Long(parser.nextText());
-                        version.setUpdateTime(time.longValue());
+                        version.setUpdateTime(time);
                     }
                     else if (parser.getName().equals("downloadURL")) {
                         version.setDownloadURL(parser.nextText());
